User uploading a file now populates basic info fields
The user uploading a file now populates the fields of the basic info page with the contents of the file.  The name field will be filling in with the file's basic_info.name attribute, etc.

This sort of works under the assumption that the user uploads a properly formatted file, but adding file validation should cover that.

For any other fields in the sheet, a programmer will need to add
value={this.props.data.ability_scores.charisma_mod}

(replace charsima mod with the corresponding name of the field in the charsheetdata.js, such as basic_info.race